Extract VRamManager to the `tiled` module
